Small cleanup; js_only.
