+++ assets/src/infrastructure/ui/inputs/Button/index.ts
@@ -1,3 +1,5 @@
 export { default as Button } from './Button';
 
+export { default as ButtonGroup } from './ButtonGroup';
+
 export { default as IconButton } from './IconButton';
