+++ assets/src/infrastructure/ui/inputs/types.ts
@@ -1,11 +1,9 @@
 import React from 'react';
 import { Textarea as ChakraTextarea } from '@chakra-ui/core';
 import type {
-	ButtonProps as ChakraButtonProps,
 	CheckboxProps as ChakraCheckboxProps,
 	CheckboxGroupProps as ChakraCheckboxGroupProps,
 	FlexProps as ChakraFlexProps,
-	IconButtonProps as ChakraIconButtonProps,
 	InputProps as ChakraInputProps,
 	NumberInputProps as ChakraNumberInputProps,
 	PseudoBoxProps as ChakraPseudoBoxProps,
@@ -19,18 +17,9 @@ export interface CommonInputProps<T = Element, V = React.ReactText> {
 	onChangeValue?: (value: V, event?: React.ChangeEvent<T> | React.FormEvent<T>) => void;
 }
 
-export interface ButtonProps extends Partial<ChakraButtonProps> {
-	buttonText?: React.ReactNode;
-	icon?: React.ComponentType<any>;
-}
-
 export interface CheckboxProps extends ChakraCheckboxProps {}
 
 export interface CheckboxGroupProps extends ChakraCheckboxGroupProps {}
-
-export interface IconButtonProps extends ChakraIconButtonProps {
-	icon: React.ComponentType<any>;
-}
 
 export interface NumberInputProps extends ChakraNumberInputProps {
 	inputFieldProps?: ChakraInputProps;
